feat(#717): serialize lifecycle flows through a mailbox
Starts already ran on a chain; resyncs and continuations now join it, so
lifecycle work runs strictly one flow at a time. Public entrypoints
queue a turn; a flow needing a sub-flow calls its inner body directly
(the resync body moves to run-resync-flow so the start flow and
continuations can), since a turn awaiting a turn queued behind itself
would deadlock. Stops stay outside for the instant halt and are observed
through the epoch, captured at request arrival so a stop landing during
the queue wait still aborts installs. Serialization makes the ownership
re-checks and the fresher-activity install guard provably dead — deleted
— and the tab-side SetActivity route goes with them: no tab sends it,
and its existence made those deletions unprovable.
