+++ libs/game/idle-client/src/index.ts
@@ -39,11 +39,9 @@ export * from './types';
 export { createInitializeMessage } from './worker/create-initialize-message';
 export { OFFLINE_CAP_WARNING_MS } from './worker/offline-cap-warning-ms';
 export { createRequestResyncMessage } from './worker/create-request-resync-message';
-export { createSetActivityMessage } from './worker/create-set-activity-message';
 export { createStartActivityMessage } from './worker/create-start-activity-message';
 export { createStopActivityMessage } from './worker/create-stop-activity-message';
 export { createSetFailureActionMessage } from './worker/create-set-failure-action-message';
 export { isRequestResyncMessage } from './worker/is-request-resync-message';
-export { isSetActivityMessage } from './worker/is-set-activity-message';
 export { isStartActivityMessage } from './worker/is-start-activity-message';
 export { useSimulationWorker } from './worker/use-simulation-worker';
