+++ libs/game/idle-client/src/test-utils/create-stub-worker-context.ts
@@ -39,7 +39,7 @@ export function createStubWorkerContext(
   let failureActionPushInFlight = false;
   let stopEpoch = 0;
   let startRequestID: null | string = null;
-  let startFlow: Readonly<Promise<void>> = Promise.resolve();
+  let lifecycleTail: Readonly<Promise<void>> = Promise.resolve();
 
   return {
     advanceStopEpoch: () => {
@@ -56,7 +56,7 @@ export function createStubWorkerContext(
       entries: rewardSlotLedger,
     }),
     getSimulation: () => simulation,
-    getStartFlow: () => startFlow,
+    getLifecycleTail: () => lifecycleTail,
     getStartRequestID: () => startRequestID,
     getStopEpoch: () => stopEpoch,
     getSubmitter: () => submitter,
@@ -98,8 +98,8 @@ export function createStubWorkerContext(
     setResyncInFlight: (inFlight) => {
       resyncInFlight = inFlight;
     },
-    setStartFlow: (flow) => {
-      startFlow = flow;
+    setLifecycleTail: (flow) => {
+      lifecycleTail = flow;
     },
     setStartRequestID: (requestID) => {
       startRequestID = requestID;
