+++ libs/game/idle-client/src/worker/handle-set-activity-message.ts
@@ -19,9 +19,9 @@ export async function handleSetActivityMessage(
     failureAction: context.getFailureAction(),
   });
 
-  // The registration starts before the activity installs but is awaited only after: installing
-  // synchronously means a resync finishing during the seed read finds this fresher activity live
-  // and yields to it, while the submitter still holds any checkpoint until the seeding resolves.
+  // The registration starts before the activity installs but is awaited only after, so the
+  // simulation is live from the first tick while the submitter still holds any checkpoint until
+  // the seeding resolves.
   const registration = context.getSubmitter().registerActivity({
     activityID: message.activity.id,
     appendedHead: message.activity.appendedHead,
