+++ libs/game/idle-client/src/worker/is-initialize-message.test.ts
@@ -1,6 +1,5 @@
 import { expect, test } from 'bun:test';
-import { createMockActivityData } from '@vers/contract-activity/test-utils';
-import type { InitializeMessage, SetActivityMessage } from '../types';
+import type { DisconnectMessage, InitializeMessage } from '../types';
 import { ClientMessageType } from '../types';
 import { isInitializeMessage } from './is-initialize-message';
 
@@ -13,9 +12,8 @@ test('it returns true if it is an initialize message', () => {
 });
 
 test('it returns false if it is not an initialize message', () => {
-  const message: SetActivityMessage = {
-    activity: createMockActivityData(),
-    type: ClientMessageType.SetActivity,
+  const message: DisconnectMessage = {
+    type: ClientMessageType.Disconnect,
   };
 
   expect(isInitializeMessage(message)).toBeFalse();
