--- libs/game/idle-client/src/worker/is-set-activity-message.test.ts
@@ -1,22 +1,0 @@
-import { expect, test } from 'bun:test';
-import { createMockActivityData } from '@vers/contract-activity/test-utils';
-import type { InitializeMessage, SetActivityMessage } from '../types';
-import { ClientMessageType } from '../types';
-import { isSetActivityMessage } from './is-set-activity-message';
-
-test('it returns true if it is a set activity message', () => {
-  const message: SetActivityMessage = {
-    activity: createMockActivityData(),
-    type: ClientMessageType.SetActivity,
-  };
-
-  expect(isSetActivityMessage(message)).toBeTrue();
-});
-
-test('it returns false if it is not a set activity message', () => {
-  const message: InitializeMessage = {
-    type: ClientMessageType.Initialize,
-  };
-
-  expect(isSetActivityMessage(message)).toBeFalse();
-});
