--- libs/game/idle-client/src/worker/is-set-activity-message.ts
@@ -1,6 +1,0 @@
-import type { ClientMessage, SetActivityMessage } from '../types';
-import { ClientMessageType } from '../types';
-
-export function isSetActivityMessage(message: ClientMessage): message is SetActivityMessage {
-  return message.type === ClientMessageType.SetActivity;
-}
