+++ libs/game/idle-client/src/worker/types.ts
@@ -58,11 +58,11 @@ export interface WorkerContext {
   readonly getSimulation: () => null | Simulation;
 
   /**
-   * The start-flow chain's tail. Start flows run one at a time — interleaved, a stale flow could
-   * stop a row the fresher one just attached. Stops stay concurrent; they bump the epoch queued
-   * flows re-check.
+   * The lifecycle mailbox's tail. Starts, resyncs, and continuations queue behind it and run one
+   * at a time — interleaved, a stale flow could stop a row a fresher one just attached. Stops
+   * stay concurrent; they bump the epoch queued flows re-check.
    */
-  readonly getStartFlow: () => Readonly<Promise<void>>;
+  readonly getLifecycleTail: () => Readonly<Promise<void>>;
 
   /**
    * The most recent start request's id. A flow that finds a fresher claim after an await abandons
@@ -106,7 +106,7 @@ export interface WorkerContext {
   readonly setFailureActionPushInFlight: (inFlight: boolean) => void;
   readonly setResyncAvatarID: (avatarID: string) => void;
   readonly setResyncInFlight: (inFlight: boolean) => void;
-  readonly setStartFlow: (flow: Readonly<Promise<void>>) => void;
+  readonly setLifecycleTail: (flow: Readonly<Promise<void>>) => void;
   readonly setStartRequestID: (requestID: string) => void;
   readonly setSimulation: (simulation: null | Simulation) => void;
 }
